% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_genres_treemap}
\alias{plot_genres_treemap}
\title{Will plot the genres present in the playlist as a treemap.}
\usage{
plot_genres_treemap(albums)
}
\arguments{
\item{albums}{A data.frame obtained with \link{request_albums}}
}
\value{
A treemap of the genres present in the playlist
}
\description{
Will plot the genres present in the playlist as a treemap.
}
