% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_timeline}
\alias{plot_timeline}
\title{Will plot the release dates of the albums present in the playlist.}
\usage{
plot_timeline(albums)
}
\arguments{
\item{albums}{a data.frame obtained with \link{request_albums}}
}
\value{
A plot of the release dates
}
\description{
Will plot the release dates of the albums present in the playlist.
}
