% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_titles_wordcloud}
\alias{plot_titles_wordcloud}
\title{Will plot a wordcloud of the words present in the tracks title}
\usage{
plot_titles_wordcloud(tracks, n = 50)
}
\arguments{
\item{tracks}{a data.frame obtained with \link{request_tracks}}

\item{n}{the maximum number of words to plot}
}
\value{
A wordcloud of top n most frequent words
}
\description{
Will plot a wordcloud of the words present in the tracks title
}
