% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_top_artists}
\alias{plot_top_artists}
\title{Will plot the top n artists (in number of songs in the playlist).}
\usage{
plot_top_artists(tracks, n = 10)
}
\arguments{
\item{tracks}{A dataframe with an "artist" column, can be obtained with \link{request_tracks}}

\item{n}{The number of artists to plot, default: 10}
}
\value{
A ggplot barplot with the top artists by number of songs
}
\description{
Will plot the top n artists (in number of songs in the playlist).
}
