% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_top_labels}
\alias{plot_top_labels}
\title{#' Will plot the distribution of the BPMs in a playlist as a beeswarm !
#'
#' @param tracks A data.frame obtained with \link{request_tracks}
#' @return a beeswarm of the BPMs
#' @export
plot_BPMs <- function(tracks) {
beeswarm(x = iris$Sepal.Length,
pch = 16,
col = "blue")
}}
\usage{
plot_top_labels(albums, n = 8)
}
\description{
#' Will plot the distribution of the BPMs in a playlist as a beeswarm !
#'
#' @param tracks A data.frame obtained with \link{request_tracks}
#' @return a beeswarm of the BPMs
#' @export
plot_BPMs <- function(tracks) {
beeswarm(x = iris$Sepal.Length,
pch = 16,
col = "blue")
}
}
