% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deezer_requests.R
\name{request_albums}
\alias{request_albums}
\title{Get the albums information:
Makes requests to Deezer's API to get the albums as a data.frame
Needed to get the genres, year of release and explicit lyrics information.}
\usage{
request_albums(tracks)
}
\arguments{
\item{tracks}{a dataframe with an "artist" column, can be obtained with
\link{request_tracks}}
}
\value{
A data.frame with the albums information
}
\description{
Get the albums information:
Makes requests to Deezer's API to get the albums as a data.frame
Needed to get the genres, year of release and explicit lyrics information.
}
