% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deezer_requests.R
\name{request_tracks}
\alias{request_tracks}
\title{Get playlist tracks:
Makes a request to Deezer's API to get the tracks as a parsed R object.}
\usage{
request_tracks(playlist_id = "6346527824")
}
\arguments{
\item{playlist_id:}{a string representing the playlist ID}
}
\value{
A data.frame with the tracks information
}
\description{
Get playlist tracks:
Makes a request to Deezer's API to get the tracks as a parsed R object.
}
\examples{
get_tracks_list("6346527824")
}
