% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deezer_requests.R
\name{request_playlist_info}
\alias{request_playlist_info}
\title{Get playlist information:
Makes a request to Deezer's API to get the tracks as a parsed R object.}
\usage{
request_playlist_info(playlist_id = "13323984883")
}
\arguments{
\item{playlist_id}{a string representing the playlist ID}
}
\value{
A parsed R object from \code{\link[httr:content]{httr::content()}}
}
\description{
Get playlist information:
Makes a request to Deezer's API to get the tracks as a parsed R object.
}
\examples{
request_playlist_info("13323984883")
}
